ENG-1391 Feat: Install component generator (#17)
## Affected Packages

- [ ] Portal
- [ ] 1UI
- [ ] Core
- [X] Automation tools

## Overview

This PR adds a separate folder called `tools`, which can house our
generators, executors, and other NX related workflow items. It contains
it's own (non published package) called `@intuition-ts/automation-tools`
- @0xjojikun @jonathanprozzi could use some ideas for a better name
perhaps.

Note that this is just an initial setup of a generator - in this case to
create a component folder for `1ui`. All other additional features and
commands can be layered on top.

Added this knowledge to the `CONTRIBUTING.md` file as well

### UI component generator

```
pnpm 1ui:component:create
```

Running the command above will ask for A name for the component is
enforced that this name should only be in PascalCase, otherwise the
command will error out. Once the name is entered, the generator creates
the scaffold files for the component inside a folder based on the
structure that was decided upon.

## Screen Captures

<img width="945" alt="image"
src="https://github.com/0xIntuition/intuition-ts/assets/37151383/3c554d20-eae7-4ac3-a6e2-c6a93ff5c0db">

## Declaration

- [X] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)