ENG-0000 Fix: Build config fixes (#23)
## Affected Packages

- [X] Portal
- [X] 1UI
- [ ] Core

## Overview

This PR addresses a few lint and build issues we had from the initial
setup. Here are few of the fixes:

- All packages now lint and build as expected
- Solves the issues from absolute imports + importing from
`@intuition-ts/1ui` etc
- Creates dist folders for both `1ui` and `core` at the root level
- Manages their builds in the project.json files so that Nx can take
care of them
- Removes unwanted setup code

Note that this unblocks us from committing and developing. We can
definitely add back anything we think is required, but this is a good
lean spot to kick on from. This will also help setup a release POC. I am
sure there will be other improvements that we can glean from that
process and iterate on this setup.

**BIG CALLOUT**

The following PostCSS config error persists (comment in Files changed).
In order to move on, I have removed the theme configuration. This
essentially means that `portal` is missing themes from `1ui`. Will
address this in a follow up PR.

![image](https://github.com/0xIntuition/intuition-ts/assets/37151383/bb853539-7aaf-471f-b16e-ad40c275d41a)


## TODO

In follow up PRs

- [ ] Fix the theme support in Portal
- [ ] Move Portal builds to Nx using project.json @nx/vite:build