ENG-1521 Feat: Adds in linking/unlinking social accounts with Privy (#52)
- Adds a `usePrivySocialLinking` hook that wraps the Privy
linking/unlinking functionality and is typesafe and reusable. Also adds
supporting Privy types
- Adds in placeholder UI focusing mainly on the functionality
- Adds the `Accordion` and `Sonner` to `1ui` -- wanted to bring these in
for the initial implementation. These aren't themed and are using the
default Shadcn styles, but figured I'd test the generator and bring it
into `1ui` directly instead of into Portal
- This is only the **client-side** implementation. Next up will be
wiring it up with our API. The hook allows for this, but will likely
need to have some additional refactoring as we implement the API calls
in the `onSuccess` (or after a successful unlinking -- there isn't an
`onSuccess` callback in the unlinks)