ENG-1599 feat(portal): add base layout and navigation (#109)
- This adds in the sidebar navigation, basic routing structure with
outlets, and a nested layout for use in routes that have the 1/3 - 2/3
layout
- The `NestedLayout` component takes in `children` and the `outlet` --
with the children being the dynamic left 1/3 and the outlet being the
outlet for the route. We can then make layout adjustments at the
component level. We won't use this nested layout on routes that don't
need it.
- Noting that this is the initial attempt and that we'll want to iterate
based on actual usage as we implement the designs for these routes and
outlets!