ENG-2111 ENG-2112 feat(portal): identity staking (#178)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

- Adds staking functionality for Identities.
- Fixes Percent Ownership on User Profile route.

NOTE: There are still a couple of components in flight in 1UI which
affect the staking modal. Once the dialog tabs and transaction state
component(s) are complete, I will revisit the staking components and
update them. As it stand, clicking Buy opens a deposit modal, and
clicking sell in the position card opens a redeem modal.

## Screen Captures


![image](https://github.com/0xIntuition/intuition-ts/assets/34991917/98020211-2f76-40d9-896f-35077724fb1f)

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)