ENG-2573 fix(portal): remove fuzzy search from explore pages (#268)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

- Added an optional boolean `isExplore` to the List and ClaimsList
components to flag whether or not it is on an explore route. When `true`
it won't render the SearchAndSort component. We could pull SearchAndSort
out of the List component, but then we'd have to import it in several
places. This seemed like the best way to handle this without duplicating
a bunch of code and having separate ExploreLists.
- Updated the search param to the respective params used in
ExploreSearch on all Explore routes.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)