ENG-2687 fix(portal): fix edit profile modal and improve ux (#296)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Made a few changes to the modal. It was exiting the modal to quickly
previously and I think that was the root cause for some details not
updating. Since handling loading and the `onClose` event more strictly,
the problem seems to be resolved. Also setting the `userObject` state
from the `offChainFetcher` response improves UX by loading the changes
what seems like instantaneously.