ENG-2714 fix(portal): nested layout sizing changes (#318)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

What started as a ticket to fix the button sizing on ClaimStakeCard
turned into adjusting the sizing of the rest of the app. The 1/3 section
(if we want to call it that) was constrained and needed be expanded
slightly (roughly ~20px). Rather than trying to squeeze those buttons
into a space that was already very tight, decided to give that entire
section some breathing room.

This also gave me further clarity into the scrolling issues we are
seeing throughout the app. Noted in `sidebar-nav.tsx`, setting
overflow-y-scroll on SidebarLayoutContent is causing issues when
navigating throughout the app. This is causing the scroll position to
not reset to the top of the page on navigation. If you scroll half way
down a page, and then navigate to another route in the sidebar, the
scroll position should reset to the top of the new page, but it doesn't.
We should discuss and figure out how to handle this. I removed some
previous "fixes" that I implemented today, because I now see the larger
issue.

- Adjusted NestedVerticalLayout, removed padding and set the width of
the content to 2/3, removed overflow class because it was rendering a
second scrollbar.
- Adjusted hardcoded widths on 1ui components that live in the 1/3
section of the app
- Removed some unnecessary divs that were inside NestedLayout

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)