ENG-2774 fix(portal): fix followClaim check on profile route (#339)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Updated the connections nested route on the profile route to match the
followClaim check used on the read-only profile route. I know that
sounds confusing.

Based on @rahulr8 comments in Linear, we should generalize this check
for these routes. I am going to work on abstracting this so that we can
use it in both routes without having to worry about keeping them in line
with eachother.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)