ENG-2797 feat(portal): explore lists (#351)
- Adds in the Explore Lists view -- realized it made sense to do this
prior to going deeper on the other Lists views
- Uses the ListGrid and a modified ListIdentityCard -- (@rahulr8 I
pulled this out of 1ui for now to make some modifications. They're
slight right now, but I'm thinking I may need to make some as I roll out
the other functionality. If it ends up being slight we can roll the
changes back into 1ui and delete the Portal specific one, provided it
doesn't include business logic)
- There are still some pieces to work out, but we can make updates to
this as we have the Lists functionality in end-to-end