ENG-2758 feat(portal): lists main layout (#365)
- Builds the `/lists` route view with the Saved and Created Tabs
- Leverages our defer/suspense pattern -- there are a few tweaks that
we'll need to make that I'm aware of
- Note: Empty State for 0 results -- I added this in but removed it
because I want to test it when we have Saved Lists in place
- Note: Search on Saved Lists is going to need to be adjusted once we
have that in place. We're currently using `displayName` here for the
query, and we can't then Search on it. I'll rethink this when we have
the actual Saved Lists to work with.