ENG-2783 feat(portal): implement defer suspense pattern on data created route (#344)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

This implements the defer and Suspense pattern from Data About to Data
Created. Couple things...

1. If the API call fails, it is throwing to the ErrorBoundary instead of
just rendering the DataErrorDisplay in place of the list.

2. We have to pass in the Skeleton component to `totalCount` on
`TabsTrigger` which we aren't able to do unless we update the type on
that component in 1ui. What do you guys think is the best path here?

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)