ENG-2791 fix(portal): swap buttons for remix links on segmented nav (#366)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Did it a little differently this time since we don't want to inject any
Remix into 1ui. Wrapped the SegmentedControlItem in a Link component so
that we can leverage prefetching. Also simplified the path logic.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)