ENG-2788 feat(portal): implement defer suspense pattern on profile connections (#346)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Had to handle this one a little differently than the others because it
relies on the `userIdentity` having a `followClaim`. Because of this I
bundled everything in one Suspense block. If the user doesn't have a
`followClaim` then it will show a string saying that, like it used to.
If it does, it will handle things the same as it does on the other
routes with this pattern.

This raises another question... do we want to handle all routes like
this? and just bundle everything under one Suspense block? or do we want
to be more granular and actually show as many UI elements as possible so
that the user feels like the app is 'working'? Open to suggestions and
feedback.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)