ENG-2819 feat(portal): stake directly on tag flow (#420)
- This leverages the Stake/Follow flow to allow users to stake directly
on existing Tags which we'll use to Save a List
- In order to trigger this from an embedded Tag it leverages 2 fetchers
-- one to get the underlying `claim_id` from the triple, and the other
to then get the `vaultDetails` from that `claim_id` -- this allows us to
trigger this pretty flexibly, but also makes it dependent on vault
details loading. I added a skeleton for this reason
- We'll likely need to make some adjustments to this flow once more
folks QA it
- Note: The screencap shows the tag being removed when I close my
position. We have a thread for this and will test the behavior once more
folks are using tags.