ENG-0000 fix(portal): Typescript issues on portal (#453)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

1. All instances of `useContractWriteAndWait` are currently being used
wrong. `functionName` doesn't seem to be part of its type interface. I
have commented it out for now but @jonathanprozzi and @Vitalsine85 you
guys will know best how to implement the real fix

2. We haad the same `TransactionStatusType` exported from 1ui but then
we were defining it in portal as well and passing it to a `1ui`
component

This PR standardizes it to use the type that is exported by 1ui in order
to use the specific status component

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)