ENG-2938 fix(portal): better ux with claim component (#462)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

This improves the UX when interacting with claims. Previously, the
entire ClaimRow linked to the claim, which limited users ability to
interact with it. Now you can hover over the IdentityTags and see a
HoverCard with a link to the IPFS document or to their wallet on the
block explorer, as well as copy the IPFS CID or wallet address.
Additionally, you can also choose to view the identity. Clicking on the
Claim component itself still navigates to the Claim Detail page as well.
This is more focused and dynamic at the same time.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)