ENG-3208 fix(portal): fix page width and add background to left column (#560)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Adds a background to the left column of detail routes that does a better
job of separating the interactive elements with the data associated with
the claim or identity. Also removes the max-7xl on the root layout, so
that the two panel layout can fill the screen. Added the necessary
padding to everything and tested on mobile to make sure it didn't break
responsiveness.

## Screen Captures


![image](https://github.com/user-attachments/assets/0b4494e9-9473-40df-ac28-ddfd7e320acd)

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)