ENG-0000 fix(portal,1ui): dialog uniformity (#563)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Made a few changes so that dialogs were a bit more uniform. Namely made
some changes to the Review and Transaction states. Now has a disabled
back button and close button in the transaction state, and all dialogs
now have a retry button in the event a transaction fails. Added a list
type to the transaction status indicator, so that it has the full copy
on success.

This also fixed an error on the create identity form where the formState
wasn't clearing and was causing the backend to throw an error saying we
were trying to make duplicate identities. Now whenever the form closes,
it clears state properly.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)