ENG-3298 feat(portal): position reconciliation maintenance tool  (#577)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

This adds a maintenance route where you can reconcile transactions
either one at a time, or by selecting a date to reconcile the
transactions from. Currently it is gated by auth, but we should probably
only allow admin users of the api, which would require BE appending role
to the user object.

## Screen Captures


![reconcile](https://github.com/user-attachments/assets/c1ec010c-5829-44be-8724-85cc181ee63b)

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)

---------

Co-authored-by: Jonathan Prozzi <jonathanprozzi@gmail.com>