ENG-0000 fix(portal): fix redeem amounts in activity feed (#595)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Fixes an issue where redeemed amounts weren't populating in activity
feed. Should probably get backend to append this field from the log to
the activity object. I've already spoken to @KeeganMyers about it but I
will raise it when things slow down a bit. Also changes the positioning
of hover cards on Claim component, so that it doesn't cover the other
atoms in the claim.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)

Co-authored-by: jojikun <110207608+0xjojikun@users.noreply.github.com>
Co-authored-by: Jonathan Prozzi <jonathanprozzi@gmail.com>