ENG-3338 fix(portal): tag identity modal hover state click on identity to route to (#609)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Adds HoverCards to the SaveForm, for when you hover over the Identity
and Tag IdentityTag. lol. There is a View Identity button in there, and
I liked how the arrows looked on JPs implementations, so I added an
arrow to this one and any others that were lacking it. Looks better and
helps to be consistent!

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)