ENG-3334 ENG-3336 fix(portal): fix crash on list detail when opening multiple save list modals (#607)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Opening more than one save list modal on the list detail page was
intermittently causing crashes when it tried to fetch things like
contract and vault_id. I made sure to clear state better onClose, added
a tag prop to TagList, and fixed an issue vite was throwing about the
contract ternary. Tested several times and unable to repro the original
issue.

This also fixes ENG-3336, as it is the same issue, just from the
identity detail route.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)