ENG-3374 fix(portal): fix initial deposit on create claim form (#628)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Fix the initial deposit not being sent on claim creation. Was seeing a
lot of scientific number still, so I updated formatNumber again and
removed the precision prop as I am just handling the formatting and
rounding internally now. Either we render out 4 decimals, or we render
out as many as we need, out to 10, otherwise render 0. Basically
anything under 1e10 will render 0 as we can consider it "dust". Users
will still see they have a position, it's just too small a number to
render on FE and the likelyhood someone has that much left in there is
not likely. It works out to like 3/1000000 of a penny. lol.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)

Co-authored-by: Jonathan Prozzi <jonathanprozzi@gmail.com>