ENG-3396 fix(portal): reduce tags in identity content row (#653)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

There were too many tags in IdentityContentRow for smaller screen sizes
and was causing some bad styling issues. Reduced the max to 2 in a
content row like that.

@jonathanprozzi I saw that in your changes yesterday you made the entire
row clickable to route to the hasTag claim. They used to be linked like
that but having a link wrapping the entire row causes issues with any
other links/clickables in the row, including the hover card. Couldn't
copy the wallet/ipfs cid without it navigating either. Don't love that
you can't click in the padding of the rows either haha. Just being nitty
at this point. Avatar/name still link to claim. I made the +X more next
to tags link to the identity, so they could view the other tags on that
identity. Feels less clunky imo and gives more functionality.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)