ENG-3389 feat(portal): handle pending states in identity details loader (#670)
- This adds in a check with our new pending routes via a
`getIdentityOrPending` call that replaces our `getIdentityById` call in
the Identity Details loader
- If an identity is a 404 it checks if it's pending. If pending, it
renders the amount of data that we have available and omits any onchain
actions and displays the refresh banner. If it's an actual 404 it'll get
caught in the 404 error boundary
- I have a comment about why I'm using the type the way we are in line
26 on `identities.ts` -- we could also set defaults for the missing
properties but I'd worry that could result in people taking actions on
incorrect data
- We can decide what we want the right panel to look like before doing
other routes -- positioning the banner, etc. -- but the logic check is
working
- Co-authored-by: Vital <rob@creative-element.net>
- Co-authored-by: Alexander Mann <alexandercmann@gmail.com>