ENG-0000 fix(portal): add timeout to create identity (#664)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Adds timeout to create identity modal. Added a little more error
handling, trying to get the API errors out of the offChainFetcher to
render better toast errors on offchain create errors. Def needs a little
more work but the basics of it are in here. Updated transaction states
with more information as well which was requested in ENG-3434

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)