ENG-3136 fix(portal): reintroduce stricter check for logout condition (#675)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Reintroduces a stricter check to logout condition. This should now
always log a user out if:
1. Their wallet is locked, regardless of if their Privy session is
valid.
2. If the wallet in the user object doesn't match the address returned
from Wagmi.
3. If address is undefined.

This might hinder DX a little initially, as a Remix server restart
causes a logout due to race condition. This is something we'll have to
deal with though, as we can't go live without this check.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)