ENG-3456 fix(portal): fix explore list search (#680)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Fix Explore Lists search. We were pruning duplicates based on
claim.object.display_name, but in the event there are two lists with the
same display_name, we are removing the other(s). Instead, we can filter
on claim.object.identity_id and it will return all the lists, even if
they have the same display name on the object. A good example currently
is the Top Web3 Developer Tooling lists.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)