ENG-0000 chore(portal): nomenclature update (#712)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

This PR updates some of the nomenclature throughout the app in favor of
uniformity. It also adds the color into the pie chart in the
PositionCard that is in the FIgma designs. Not sure why that was never
implemeneted, looks great.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)