ENG-0000 fix(portal): fix home loop (#751)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Fixes the loop when logging in where users were sent to Home just to be
redirected back to log in. Forgot to move the redirect login from
Profile to Home when we switched the rest of the redirect logic to now
bring users to Home on login. Added a better check at the app layer. Can
improve on this further but going to save that for post launch.