ENG-3607 fix(portal): update min deposit amounts in radio selects (#750)
- Updated min deposit on radio select forms (following, save list). Added
a min_deposit const as a fallback if vaultDetails doesn't resolve for
some reason. Moved the SaveListModal into the route instead of there
being an instance in every one of the rows in the list. Can probably
move all modals to the app root so we don't have multiple instances of
them and they're handle with Jotai.
- Co-authored-by: Jonathan Prozzi <jonathanprozzi@gmail.com>