ENG-0000 fix(portal): fix staking modal bugs (#770)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Changes got overwritten somehow that set values based on whether it was
an identity or claim in the modal. Because of this position and redeem
functionality was busted. Fixed the problem by passing the appropriate
values down from the modal.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)