ENG-0000 fix(portal): remove consoles and fix max button math (#775)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Removes consoles. Fixes funds check for deposits. Not sure why it was
multiplying val by conviction price, cause that's not correct. We are
inputting ETH and the contract expects ETH, so we should only be
comparing val to walletBalance. Also includes a fix for max buttons so
that they don't put scientific numbers.