ENG-3737 fix(portal): add skip button to onboarding videos (#800)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Added a skip button to the relic onboarding videos. Also abstracted the
video into a component that we can reuse throughout the flow.

Also did some testing for ENG-3744 and was unable to repro the issue. No
changes were made related to this.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)