ENG-3797 ENG-3736 fix(portal): lists and claims back button and hook (#805)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

- Refines the back button approach to go backwards through the history
stack, including query params in the URL
- Handles case where a user navigates via a fresh link to the route (no
history) and navigates to a fallback. This is a bit slower due to the
way we need to handle this -- if we don't care about this (default is
goes back to a blank tab) we can remove this part of the hook. However,
this allows us to keep users within the app when they click back on a
fresh route
- Creates a `useGoBack` hook to manage this logic


## Screen Captures

In app navigation with query params:

![in-app-nav-back-button](https://github.com/user-attachments/assets/1a401052-30f5-48c7-92e4-5c6888a62c56)

Landing on a page for the first time, such a from a link:

![from-link-back-button](https://github.com/user-attachments/assets/19be510b-0114-45af-b867-6a07561208c4)

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)

 
 **PR Summary by Typo**
------------

 **Summary:**
- New constant `EXPLORE_CLAIMS` added to `paths.ts` and new hooks
`useGoBack` and state atoms `addIdentitiesListModalAtom`,
`imageModalAtom` imported and used.
- `handleGoBack` function added to `ClaimDetails.tsx` using `useGoBack`
hook, and `NavigationButton` component modified to use `handleGoBack`
and an `onClick` event listener.

**Key Points:**
- In `paths.ts`, a new constant `EXPLORE_CLAIMS` is added.
- In `useGoBack.tsx`, a new custom hook `useGoBack` is created, and in
`ClaimDetails.tsx`, the `handleGoBack` function is added using this
hook.
- In `ListDetails.tsx`, `useGoBack`, `useLiveLoader`,
`addIdentitiesListModalAtom`, and `imageModalAtom` hooks are imported
and used. The `NavigationButton` component is modified to use the
`handleGoBack` function and an `onClick` event listener.

**Categories:**
- Constants: `paths.ts`
- Hooks: `useGoBack.tsx`, `useLiveLoader.tsx`
- State Management: `addIdentitiesListModalAtom.ts`, `imageModalAtom.ts`
- Components: `ClaimDetails.tsx`, `ListDetails.tsx`
- Files: `apps/portal/app/consts/paths.ts`,
`apps/portal/app/lib/hooks/useGoBack.tsx`,
`apps/portal/app/lib/hooks/useLiveLoader.tsx`,
`apps/portal/app/routes/app+/claim+/$id.tsx`,
`apps/portal/app/components/ListDetails.tsx`

<h6>To turn off PR summary, please visit <a
href="https://app.typoapp.io/settings/dev-analytics/notification?tab=codeHealth">Notification
settings</a>.</h6>