ENG-3826 fix(portal): fix video overflow on create (#813)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

The relic video after creating your user identity was still overflowing
after the last fix. Forgot to apply the same styling to the parent
containers which should constrain it to the size of the screen. You have
to zoom in over 400% for it to overflow now, and it looks good on
mobile. Should be good from here.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)

 
 **PR Summary by Typo**
------------

 **Summary**
This pull request modifies the layout of the `Profile` component in
`create.tsx`.

**Key Points**
1. Adjusted the height of the main content area and padding.
2. Changed the inner `div` structure.

**Layout adjustments**
- **Min-height adjustment:** Changed the height of the main content area
from `h-screen` to `min-h-screen`.
- **Padding adjustment:** Removed the fixed height for the padding in
smaller screens (`md:p-8`).
- **Inner div structure:** Changed the class of the inner `div` from
`flex justify-center items-center h-screen` to `flex-grow flex
justify-center items-center`.

<h6>To turn off PR summary, please visit <a
href="https://app.typoapp.io/settings/dev-analytics/notification?tab=codeHealth">Notification
settings</a>.</h6>