ENG-3539 feat(1ui,portal): Row hover UX improvements (#810)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

First, I probably could have held off, but it was kind of messy in the
/components/list folder, there were components for paginated lists, as
well as lists related components (confusing I know lol), and components
for the save list modal. I moved the save list modal components to their
own folder, and the lists related components to their own folder. Maybe
I should have done it in a separate PR, I can roll it back and do it in
another, if you guys want. It was just easier for me to work through the
lists after cleaning that folder up.

It was requested that we make navigation more intuitive when it comes to
interacting paginated lists, specifically claims. Let's focus on the
changes to claims first:

Previously, you could click anywhere inside of the Claim component to
navigate to the claim detail route. Now, clicking anywhere inside of the
ClaimRow will navigate to the claim detail. I also added a hover effect
to make it more obvious that the row is clickable. Hovering over the
identities in the claim has the same effect as before, rendering a hover
card, but now clicking on the identities themselves inside of the claim,
will navigate to the identity detail page. When hovering over the
ClaimRow, all of the IdentityTags that comprise the Claim component
highlight, but when you hover over just one of them, only that one his
highlighted, to help establish that the user is interacting with that
individual claim.

It was a bit tricky to get this working, as everything is nested inside
of the ClaimRow, and wrapping that in any type of anchor prevents us
from being able to interact with any elements inside of that ClaimRow,
and there are quite a few.

I applied these improvements to all other row components as well, so
that the hover styling is consistent, and it gives users more clickable
area to navigate on the explore routes, and paginated lists in general.

I think that these changes will improve UX until design has had some
time to sit down and rethink the flow in general for explore pages and
paginated lists in general.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)

 
 **PR Summary by Typo**
------------

 **Summary:**
This pull request introduces a new component `ClaimPositionRow` with
hooks for handling clicks, mouse events, and keyboard events on link or
interactive elements. It also includes imports and types for the new
functionality.

**Key Points:**
- A new component `ClaimPositionRow` was added with file extension
`.tsx`.
- The component uses hooks for state and refs, and imports necessary
dependencies.
- The component accepts various props such as position, amount,
currency, feesAccrued, updatedAt, tags, link, and ipfsLink.
- The component has two variants, 'user' and 'claim', with different
prop structures.

**Categories:**

**Components:**
- Added a new component `ClaimPositionRow` to the
`claim-position-row.tsx` file.

**Hooks:**
- The component uses `useState` hook for handling interactive element
state.
- The component uses `useRef` hook for handling link ref.

**Dependencies:**
- Imported various dependencies such as 'Avatar', 'Button',
'ClaimPositionType', 'ClaimStatus', 'cn', 'Copy', 'CurrencyType',
'formatDate', 'HoverCard', 'HoverCardContent', 'HoverCardTrigger',
'Icon', 'PositionValueDisplay', 'ProfileCard', 'TagsContent',
'TagWithValue', 'Text', 'TextVariant', 'TextWeight', 'Trunctacular',
'@remix-run/react', 'useNavigate', 'ClaimPositionRowVariant',
'ClaimPositionRowVariantType',
'@0xintuition/1ui/src/components/ClaimPositionRow/ClaimPositionRow.utils'.

**Types:**
- Defined interfaces 'CommonProps', 'UserVariantProps',
'ClaimVariantProps', and 'ClaimPositionRowProps'.
- Defined type 'ClaimPositionRowProps' as a union of 'UserVariantProps'
and 'ClaimVariantProps'.

**Rendering:**
- Removed `role` and `tabIndex` attributes when `isInteractiveElement`
is false.
- Added dynamic classes based on `isInteractiveElement` and `variant`
props.
- Conditionally rendered a `HoverCard` component with user information
when `variant` is `ClaimPositionRowVariant.user`.
- Added hover effects and a button for viewing identity in the
`HoverCardContent`.
- Added truncated text for name and ID with copy functionality.
- Displayed last update date if `updatedAt` is present.

**Events:**
- Added event handlers for onClick, onKeyDown, and onMouseMove.
- Implemented state variables and refs to manage interactive elements
and navigation.

**Accessibility:**
- Changed role and tabIndex based on the state of
`isInteractiveElement`.

<h6>To turn off PR summary, please visit <a
href="https://app.typoapp.io/settings/dev-analytics/notification?tab=codeHealth">Notification
settings</a>.</h6>